+++ src/components/ColorStop/types.ts
@@ -1,23 +1,11 @@
 import { RefObject } from 'react';
+import { ColorStop, Limits, DragCoordinates } from '../../types';
 import { DIRECTIONS } from '../GradientPicker/constants';
 
-export interface Stop {
-  id: number;
-  color: string;
-  offset: number;
-  isActive: boolean;
-  opacity?: number;
-  pointX?: number;
-}
-
-export interface Limits {
-  min: number;
-  max: number;
-  drop?: number;
-}
+export type { DragCoordinates };
 
 export interface ColorStopProps {
-  stop: Stop;
+  stop: ColorStop;
   limits: Limits;
   onPosChange: (data: { id: number; offset: number }) => void;
   onDeleteColor: (id: number) => void;
@@ -27,7 +15,7 @@ export interface ColorStopProps {
 }
 
 export interface UseStopDraggingProps {
-  stop: Stop;
+  stop: ColorStop;
   limits: Limits;
   initialPos?: number;
   colorStopRef: RefObject<HTMLDivElement | null>;
@@ -36,9 +24,4 @@ export interface UseStopDraggingProps {
   onDragEnd: (id: number) => void;
   onDeleteColor: (id: number) => void;
   direction: (typeof DIRECTIONS)[keyof typeof DIRECTIONS];
-}
-
-export interface DragCoordinates {
-  clientX: number;
-  clientY: number;
 }
