+++ src/components/GradientTypePicker/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { GRADIENT_TYPES, GradientTypePickerProps } from './types';
+import { GRADIENT_TYPES } from '../../types';
+import { GradientTypePickerProps } from './types';
 import './index.scss';
 
 const GradientTypePicker: React.FC<GradientTypePickerProps> = ({ gradientType, onGradientTypeChange }) => {
