+++ src/types/geometry.ts
@@ -1,0 +1,11 @@
+export interface Point {
+  x: number;
+  y: number;
+}
+
+export interface GradientCoordinates {
+  x1: number;
+  y1: number;
+  x2: number;
+  y2: number;
+}
