+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './color';
+export * from './common';
+export * from './geometry';
+export * from './gradient';
